pyproject.toml
[project]
name = "selenium-course"
version = "0.1.0"
description = ""
authors = [
    {name = "s-sorokaty",email = "sorokaty0824@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.32.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
